+++ post-json.js
@@ -1,0 +1,17 @@
+const http = require("http");
+
+const host = "localhost";
+const port = 8000;
+
+// // POST /json
+const requestListener = function (req, res) {
+  res.setHeader("Content-Type", "application/json");
+  res.writeHead(200);
+  res.end(`{"message": "This is a JSON Response"}`);
+};
+// curl http://localhost:8000
+
+const server = http.createServer(requestListener);
+server.listen(port, host, () => {
+  console.log(`Server is running on http://${host}:${port}`);
+});
